# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for AI Agent Development Guidebook
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search:
    get:
      operationId: webSearch
      tags: [search]
      summary: Perform a web search via DuckDuckGo (no API key required)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebSearchResponse"

  # ─── Tool Evaluation ──────────────────────────────────────────────────────────
  /tool-evaluations:
    get:
      operationId: listToolEvaluations
      tags: [tool-evaluation]
      summary: List all tool evaluations
      responses:
        "200":
          description: List of evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolEvaluation"
    post:
      operationId: createToolEvaluation
      tags: [tool-evaluation]
      summary: Create and run a new tool evaluation (SSE streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolEvaluationBody"
      responses:
        "200":
          description: SSE stream of evaluation progress
  /tool-evaluations/{id}:
    get:
      operationId: getToolEvaluation
      tags: [tool-evaluation]
      summary: Get a completed tool evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tool evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolEvaluation"
    delete:
      operationId: deleteToolEvaluation
      tags: [tool-evaluation]
      summary: Delete a tool evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Memory ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ToolEvaluation:
      type: object
      properties:
        id:
          type: integer
        toolName:
          type: string
        toolUrl:
          type: ["string", "null"]
        projectRequirements:
          type: string
        report:
          type: ["string", "null"]
        fitScore:
          type: ["integer", "null"]
        recommendation:
          type: ["string", "null"]
          enum: [USE, USE_WITH_CAVEATS, AVOID, null]
        status:
          type: string
          enum: [pending, completed, failed]
        language:
          type: string
          enum: [ar, en]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - toolName
        - projectRequirements
        - status
        - language
        - createdAt
        - updatedAt
    CreateToolEvaluationBody:
      type: object
      properties:
        toolName:
          type: string
        toolUrl:
          type: ["string", "null"]
        projectRequirements:
          type: string
        language:
          type: string
          enum: [ar, en]
      required:
        - toolName
        - projectRequirements

    # ─── Memory ─────────────────────────────────────────────────────────────────
...
    WebSearchResult:
      type: object
      properties:
        title:
          type: string
        snippet:
          type: string
        url:
          type: string
      required:
        - title
        - snippet
        - url
    WebSearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/WebSearchResult"
        total:
          type: integer
      required:
        - query
        - results
        - total